Ability to add different storage engines, huge refactoring
